requirements.txt
Django>=3.1.7,<3.2.0
pylint>=2.7.4,<2.8.0
pylint-django>=2.4.2,<2.5.0
djangorestframework>=3.12.4,<3.13.0
Markdown>=3.3.4,<3.4.0
django-filter>=2.4.0,<3.4.0
psycopg2-binary>=2.8.6,<2.9.0
django-environ>=0.4.5,<1.0.0
asgiref==3.2.10
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1